+++ packages/stripe_platform_interface/lib/src/models/card_field_input.dart
@@ -2,6 +2,8 @@ import 'package:flutter/painting.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:json_annotation/json_annotation.dart';
 
+import 'color.dart';
+
 part 'card_field_input.freezed.dart';
 part 'card_field_input.g.dart';
 
@@ -126,24 +128,6 @@ class CardPlaceholder with _$CardPlaceholder {
         cvc: placeholder?.cvc ?? cvc,
         postalCode: placeholder?.postalCode ?? postalCode,
       );
-}
-
-/// Deserialization object for colors.
-///
-/// This object is used to translate Flutter color objects to hex strings used by the stripe sdk.
-class ColorKey {
-  const ColorKey();
-
-  static String? toJson(Color? color) {
-    if (color != null) {
-      // ignore: lines_longer_than_80_chars
-      return '#${color.value.toRadixString(16).padLeft(8, '0').toUpperCase()}';
-    }
-  }
-
-  static Color? fromJson(value) {
-    throw UnimplementedError();
-  }
 }
 
 @freezed